+++ protobuf/zwift_messages.proto
@@ -1,0 +1,285 @@
+syntax="proto3";
+package ZwiftPacketMonitor;
+
+message Meetup {
+    int32 id = 1;
+    int32 tag2 = 2;
+    string name = 3;
+    string description = 4;
+    int64 start_time = 5;
+    fixed32 distance = 7;
+    int32 tag8 = 8;
+
+    message MeetupInner {
+        int32 id = 1;
+        string name = 2;
+        string description = 3;
+        string f7 = 7;
+        int64 f8 = 8;
+        int64 f9 = 9;
+        int64 f10 = 10;
+        int64 f11 = 11;
+        int64 f12 = 12;
+        int64 f13 = 13;
+        int64 f14 = 14;
+        int64 f15 = 15;
+        int64 f16 = 16;
+        int64 f17 = 17;
+        int64 f18 = 18;
+        int64 f22 = 22;
+        fixed32 distance = 24;
+        int64 f25 = 25;
+        int32 f29 = 29;
+        int32 f30 = 30;
+        int32 f31 = 31;
+        fixed32 f32 = 32;
+        fixed32 f33 = 33;
+        int32 duration = 34;
+        int64 f42 = 42;
+        int64 f44 = 44;
+        int32 f46 = 46;
+        int32 f47 = 47;
+    }
+    MeetupInner meetup_inner = 10;
+    string thumbnail = 12;
+    int32 duration = 13;
+    int64 f14 = 14;
+    int32 f17 = 17;
+    int32 f18 = 18;
+    int32 f22 = 22;
+    int32 f23 = 23;
+    int32 meetup_only_view = 24;
+    int32 f25 = 25;
+    int32 f27 = 27;
+    int32 f28 = 28;
+    int32 f29 = 29;
+    int32 f34 = 34;
+    int32 f35 = 35;   
+}
+
+message PlayerState {
+    int32 id = 1;
+    int64 worldTime = 2;
+    int32 distance = 3;
+    int32 roadTime = 4;
+    int32 laps = 5;
+    int32 speed = 6;
+    int32 roadPosition = 8;
+    int32 cadenceUHz = 9;
+    int32 heartrate = 11;
+    int32 power = 12;
+    int64 heading = 13;
+    int32 lean = 14;
+    int32 climbing = 15;
+    int32 time = 16;
+    int32 f19 = 19;
+    int32 f20 = 20;
+    int32 progress = 21;
+    int64 customisationId = 22;
+    int32 justWatching = 23;
+    int32 calories = 24;
+    float x = 25;
+    float altitude = 26;
+    float y = 27;
+    int32 watchingRiderId = 28;
+    int32 groupId = 29;
+    int64 sport = 31;
+    float f34 = 34;  // actual distance moved included lateral movement
+}
+
+message ClientToServer {
+    int32 connected = 1;
+    int32 rider_id = 2;
+    int64 world_time = 3;
+    int32 seqno = 4;
+    PlayerState state = 7;
+    int64 tag8 = 8;
+    int64 tag9 = 9;
+    int64 last_update = 10;
+    int64 tag11 = 11;
+    int64 last_player_update = 12;
+}
+
+message UnknownMessage1 {
+    // string firstName=7;
+    // string lastName=8;
+    // string timestamp=17;
+}
+
+message UnknownMessage {
+    // int64 tag1=1;
+    // UnknownMessage1 tag4=4;
+}
+
+
+message PlayerUpdate {
+    int64 tag1 = 1;
+    int32 tag2 = 2;
+    int32 tag3 = 3;
+    bytes payload = 4;
+    int64 tag5 = 5;
+    int64 tag6 = 6;
+    int64 tag7 = 7;
+    int64 tag8 = 8;
+    int64 tag9 = 9;
+    int64 tag11 = 11;
+    int64 tag12 = 12;
+    int64 tag14 = 14;
+    int64 tag15 = 15;
+}
+
+message Payload105 { // player entered world ?
+    int64 f1 = 1;
+    int32 rider_id = 2;
+    int32 f3 = 3;
+    int64 f4 = 4; // int32?
+    int64 f5 = 5; 
+    int64 f6 = 6; // int32?
+    string firstName = 7;
+    string lastName = 8;
+    int64 f9 = 9;
+    int64 f11 = 11;
+    int32 f12 = 12;
+    int32 f13 = 13;
+    int32 f14 = 14;
+    int32 f15 = 15;
+    int32 f16 = 16;
+    string f7date = 17;
+    int32 f19 = 19;
+}
+
+message Chat {
+    int32 rider_id = 1;
+    int32 to_rider_id = 2; // 0 if public message
+    int32 f3 = 3; // always value 1 ?
+    string firstName = 4;
+    string lastName = 5;
+    string message = 6;
+    string avatar = 7;
+    int32 countryCode = 8;
+    int32 eventSubgroup = 11;
+}
+
+message RideOn {
+    int32 rider_id = 1;
+    int32 to_rider_id = 2;
+    string firstName = 3;
+    string lastName = 4;
+    int32 countryCode = 5;
+}
+
+message Payload2 {
+    int32 f1 = 1;
+    int64 f2 = 2;
+}
+
+message TimeSync { 
+    int32 rider_id = 1;
+    int64 world_time = 2;
+    int32 f3 = 3;
+}
+
+message Payload110 { 
+    // format to be determined
+}
+
+message Payload109 { 
+    // format to be determined
+}
+
+
+message EventPositions {
+    int32 position = 1;
+    message EventRiderPosition {
+        int32 rider_id = 1;
+        // ?? float distance_covered = 2; 
+    }
+    repeated EventRiderPosition eventRiderPosition = 4;
+    int32 num_riders = 116;
+}
+
+message ServerToClient {
+    int32 tag1 = 1;
+    int32 rider_id = 2;
+    int64 world_time = 3;
+    int32 seqno = 4;
+    repeated PlayerState player_states = 8;
+    repeated PlayerUpdate player_updates = 9;
+    int64 tag11 = 11;
+    string local_ip = 12;
+    int32 tag15 = 15;
+    int64 tag17 = 17;
+    int32 num_msgs = 18;
+    int32 msgnum = 19;
+    EventPositions event_positions = 23;
+    Server1 servers1 = 24;
+    Server2 servers2 = 25;
+}
+
+message ServerAddress {
+    int32 f1 = 1;
+    int32 f2 = 2;
+    string ip = 3;
+    int32 f4 = 4;
+    fixed32 f5 = 5;
+    fixed32 f6 = 6;
+}
+
+message ServerPool {
+    int32 f1 = 1;
+    int32 f2 = 2;
+    repeated ServerAddress addresses = 3;
+    int32 f4 = 4;
+}
+
+message Server2 {
+    repeated ServerPool pool = 1;
+    int32 f2 = 2;
+}
+
+message Server1 {
+    repeated ServerAddress addresses = 1;
+    int32 f2 = 2;
+    int32 f3 = 3;
+    int32 f4 = 4;
+}
+
+message WorldAttributes {
+    int32 world_id = 1;
+    string name = 2;
+    int64 tag3 = 3;
+    int64 tag5 = 4;
+    int64 world_time = 6;
+    int64 clock_time = 7;
+}
+
+message WorldAttribute {
+    int64 world_time = 2;
+}
+
+message EventSubgroupProtobuf {
+    int32 id = 1;
+    string name = 2;
+    int32 rules = 8;
+    int32 route = 22;
+    int32 laps = 25;
+    int32 startLocation = 29;
+    int32 label = 30;
+    int32 paceType = 31;
+    int32 jerseyHash = 36;
+}
+
+message RiderAttributes {
+    int32 f2 = 2;
+    int32 f3 = 3;
+    message AttributeMessage {
+        int32 myId = 1;
+        int32 theirId = 2;
+        string firstName = 3;
+        string lastName = 4;
+        int32 countryCode = 5;
+    }
+    AttributeMessage attributeMessage = 4;
+    int32 theirId = 10;
+    int32 f13 = 13;
+}
