--- requirement.txt
@@ -1,24 +1,0 @@
-certifi==2021.10.8
-cffi==1.15.0
-chardet==4.0.0
-click==8.0.3
-colorama==0.4.4
-cryptography==36.0.1
-dnspython==2.1.0
-Flask==2.0.2
-future==0.18.2
-itsdangerous==2.0.1
-Jinja2==3.0.3
-ldap3==2.9.1
-ldapdomaindump==0.9.3
-MarkupSafe==2.0.1
-pcapy==0.11.4
-protobuf==3.19.1
-pyasn1==0.4.8
-pycparser==2.21
-pycryptodomex==3.12.0
-pyOpenSSL==21.0.0
-scapy==2.4.5
-six==1.16.0
-Werkzeug==2.0.2
-wincertstore==0.2
