+++ setup.cfg
@@ -1,9 +1,9 @@
 # python -m build
 [metadata]
-name = zwifty-packets
+name = zwiftpktmon
 version = 0.0.1
-author = Example Author
-author_email = author@example.com
+author = Kevin Ruff
+author_email = ruff.kevin@outlook.com
 description = A small example package
 long_description = file: README.md
 long_description_content_type = text/markdown
@@ -14,15 +14,25 @@ classifiers =
     Programming Language :: Python :: 3
     License :: OSI Approved :: MIT License
     Operating System :: Microsoft :: Windows
+    Environment :: MacOS X
+    Environment :: Win32 (MS Windows)
+    Intended Audience :: Developers
+    Operating System :: MacOS
+    Operating System :: Microsoft :: Windows :: Windows 10
 
 [options]
 package_dir =
-    = zpm
+    = src
 packages = find:
+
 python_requires = >=3.10
 
 install_requires = 
     protobuf >= 3.19.4
+    pcapy-ng-win-amd64 >= 1.0.4; platform_system == "Windows"
+    pcapy-ng-macos >= 1.0.4; platform_system == "Darwin"
 
 [options.packages.find]
-where = zpm
+where = src
+include = zwiftpktmon
+
