+++ tests/context.py
@@ -1,4 +1,3 @@
-
 """
 This file helps the scripts in the tests folder to find the required modules.
 """
@@ -6,4 +5,5 @@ import os
 import sys
 # Assure the project root is in the path
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
-#print(sys.path)
+
+# print(sys.path)
