--- zpm/__init__.py
@@ -1,11 +1,0 @@
-from .async_monitor import Monitor
-from .events import RiderStateEventArgs, RiderEnteredWorldEventArgs, RiderTimeSyncEventArgs, IncomingMeetupEventArgs, \
-    RiderChatEventArgs, RiderRideOnEventArgs, ZwiftEvent, ZwiftEventArgs
-
-__all__ = (
-    "Monitor", 
-    "RiderStateEventArgs", "RiderEnteredWorldEventArgs", "RiderTimeSyncEventArgs", "IncomingMeetupEventArgs",
-    "RiderChatEventArgs", "RiderRideOnEventArgs", "ZwiftEvent", "ZwiftEventArgs"
-)
-
-#print(f"__init__.py (zpm) {dir()}")
