--- zpm/protobuf/__init__.py
@@ -1,8 +1,0 @@
-from zpm.protobuf.zwift_messages_pb2 import ServerToClient, ClientToServer
-
-__all__ = (
-    "ServerToClient", "ClientToServer"
-)
-
-#print(f"__init__.py (zpm.protobuf) {dir()}")
-
