--- zpm/protobuf/previous/zwift_messages.proto
@@ -1,251 +1,0 @@
-syntax="proto3";
-
-message PlayerState {
-    int32 id = 1;
-    int64 worldTime = 2;
-    int32 distance = 3;
-    int32 roadTime = 4;
-    int32 laps = 5;
-    int32 speed = 6;
-    int32 roadPosition = 8;
-    int32 cadenceUHz = 9;
-    int32 heartrate = 11;
-    int32 power = 12;
-    int64 heading = 13;
-    int32 lean = 14;
-    int32 climbing = 15;
-    int32 time = 16;
-    int32 f19 = 19;
-    int32 f20 = 20;
-    int32 progress = 21;
-    int64 customisationId = 22;
-    int32 justWatching = 23;
-    int32 calories = 24;
-    float x = 25;
-    float altitude = 26;
-    float y = 27;
-    int32 watchingRiderId = 28;
-    int32 groupId = 29;
-    int64 sport = 31;
-}
-
-message ClientToServer {
-    int32 connected = 1;
-    int32 rider_id = 2;
-    int64 world_time = 3;
-    PlayerState state = 7;
-    int32 seqno = 4;
-    int64 tag8 = 8;
-    int64 tag9 = 9;
-    int64 last_update = 10;
-    int64 tag11 = 11;
-    int64 last_player_update = 12;
-}
-
-message SegmentResult {
-    int64 id = 1;
-    int64 rider_id = 2;
-    int64 event_subgroup_id = 6;
-    string first_name = 7;
-    string last_name = 8;
-    string finish_time_str = 10;
-    int64 elapsed_ms = 11;
-    int32 powermeter = 12;
-    int32 weight = 13;
-    int32 power = 15;
-    int32 heartrate = 19;
-}
-
-message SegmentResults {
-    int64 world_id = 1;
-    int64 segment_id = 2;
-    int64 event_subgroup_id = 3;
-    repeated SegmentResult segment_results = 4;
-}
-
-message UnknownMessage1 {
-    // string firstName=7;
-    // string lastName=8;
-    // string timestamp=17;
-}
-
-message UnknownMessage {
-    // int64 tag1=1;
-    // UnknownMessage1 tag4=4;
-}
-
-message ServerToClient {
-    int32 tag1 = 1;
-    int32 rider_id = 2;
-    int64 world_time = 3;
-    int32 seqno = 4;
-    repeated PlayerState player_states = 8;
-    repeated UnknownMessage player_updates = 9;
-    int64 tag11 = 11;
-    int64 tag17 = 17;
-    int32 num_msgs = 18;
-    int32 msgnum = 19;
-}
-
-message WorldAttributes {
-    int32 world_id = 1;
-    string name = 2;
-    int64 tag3 = 3;
-    int64 tag5 = 4;
-    int64 world_time = 6;
-    int64 clock_time = 7;
-}
-
-message WorldAttribute {
-    int64 world_time = 2;
-}
-
-message EventSubgroupProtobuf {
-    int32 id = 1;
-    string name = 2;
-    int32 rules = 8;
-    int32 route = 22;
-    int32 laps = 25;
-    int32 startLocation = 29;
-    int32 label = 30;
-    int32 paceType = 31;
-    int32 jerseyHash = 36;
-}
-
-message RiderAttributes {
-    int32 f2 = 2;
-    int32 f3 = 3;
-    message AttributeMessage {
-        int32 myId = 1;
-        int32 theirId = 2;
-        string firstName = 3;
-        string lastName = 4;
-        int32 countryCode = 5;
-    }
-    AttributeMessage attributeMessage = 4;
-    int32 theirId = 10;
-    int32 f13 = 13;
-}
-
-message Profiles {
-    repeated Profile profiles = 1;
-}
-
-message Profile {
-	int32 id = 1;
-	string firstName = 4;
-	string lastName = 5;
-	int32 male = 6;
-	int32 weight = 9;
-	int32 bodyType = 12;
-	int32 countryCode = 34;
-	int32 totalDistance = 35;
-	int32 totalDistanceClimbed = 36;
-	int32 totalTimeInMinutes = 37;
-	int32 totalWattHours = 41;
-	int32 height = 42;
-	int32 totalExperiencePoints = 46;
-	int32 achievementLevel = 49;
-	int32 powerSource = 52;
-	int32 age = 55;
-	string launchedGameClient = 108;
-	int32 currentActivityId = 109;
-}
-
-// ZML communication
-
-//  ZML server side
-
-message Vector3 {
-    float x = 1;
-    float y = 2;
-    float z = 3;
-}
-message PlayerInfo {
-    int32 id = 1;
-    int32 f2 = 2;
-    Vector3 position = 3;
-    string profile = 5;
-    int32 id2 = 6;
-    int32 f7 = 7;
-    string name = 11;
-    int32 countryCode = 12;
-    fixed32 worldTime = 13;
-    int32 f16 = 16;
-}
-
-message GTPC21_6_1 {
-    int32 seqno = 1;
-    repeated PlayerInfo playerInfos = 2;
-    int32 f3 = 3;
-}
-
-message GTPC21_6 {
-    repeated GTPC21_6_1 gtpc21_6_1 = 1;
-}
-
-message GTPC21_4 {
-    int32 f1 = 1;
-    string f6 = 6;
-    int32 f7 = 7;
-    int32 f8 = 8;
-}
-
-message GTPC21_8 {
-    int32 f1 = 1;
-    int32 f2 = 2;
-}
-
-message GTPC21 {
-    int32 f1 = 1;
-    GTPC21_4 gtpc21_4 = 4;
-    GTPC21_6 gtpc21_6 = 6;
-    GTPC21_8 gtpc21_8 = 8;
-}
-
-message GameToPhoneCommand {
-    int32 seqno = 1;
-    int32 f2 = 2;
-    GTPC21 gtpc21 = 21;
-}
-
-message GameToPhone {
-    int32 f1 = 1;
-    int32 f2 = 2;
-    int32 id = 3;
-    int32 f4 = 4;
-    int32 f6 = 6;
-    int32 f7 = 7;
-    repeated GameToPhoneCommand commands = 11;
-}
-
-//  ZML client side
-
-message ZMLClientInfo {
-    string appVersion = 1;
-    string systemOSVersion = 2;
-    string systemOS = 3;
-    string systemHardware = 4;
-}
-
-message ZMLClientCapabilities {
-    int32 f1 = 1;
-    ZMLClientInfo info = 5;
-}
-
-message PhoneToGameCommand {
-    int32 seqno = 1;
-    int32 command = 2;
-    int32 subject = 3;
-    int32 f5 = 5;
-    string f6 = 6;
-    int32 f7 = 7;
-    int32 playerId = 19;
-    ZMLClientCapabilities capabilities = 21;
-}
-
-message PhoneToGame {
-    int32 id = 1;
-    PhoneToGameCommand command = 2;
-    int32 f10 = 10;
-}
